Rename package to opennav_following
