+++ opennav_following/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>opennav_dynamic_following</name>
+  <name>opennav_following</name>
   <version>0.0.1</version>
   <description>A Task Server for dynamic following object</description>
   <maintainer email="steve@opennav.org">Steve Macenski</maintainer>
